Validation and error handling

+++ controllers/contacts.controller.js
@@ -3,17 +3,22 @@ const ObjectId = require('mongodb').ObjectId;
 
 const getAllContacts = async (req, res) => {
   const result = await mongodb.getDatabase().db().collection('contacts').find();
-  result.toArray().then(users => {
+  result.toArray().then((users) => {
     return res.status(200).json(users)
-  });
+  }).catch(error=> res.status(400).json(error));
 }
 
 const getContact = async (req, res) => {
+  // id validation
+  if(!ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({error: 'invalid id'})
+  }
+
   const contactId = new ObjectId(req.params.id);
-  const result = await mongodb.getDatabase().db().collection('contacts').find({_id: contactId});
-  result.toArray().then(users => {
+  const result = await mongodb.getDatabase().db().collection('contacts').find({_id: contactId})
+  result.toArray().then((users) => {
     return res.status(200).json(users[0])
-  }).catch(error=>console.log(error));
+  }).catch(error=> res.status(400).json(error))
 }
 
 const createContact = async (req, res) => {
@@ -25,6 +30,11 @@ const createContact = async (req, res) => {
 }
 
 const updateContact = async (req, res) => {
+  // id validation
+  if(!ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({error: 'invalid id'})
+  }
+
   const {firstName, lastName, email, favoriteColor, birthday} = req.body;
   const contactId = {_id: new ObjectId(req.params.id)};
   const updateDoc = {$set: {firstName, lastName, email, favoriteColor, birthday}};
@@ -33,12 +43,17 @@ const updateContact = async (req, res) => {
 }
 
 const deleteContact = async (req, res) => {
+  // id validation
+  if(!ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({error: 'invalid id'})
+  }
+
   const contactId = {_id: new ObjectId(req.params.id)};
   const result = await mongodb.getDatabase().db().collection('contacts').deleteOne(contactId);
   if (result.deletedCount === 1) {
-    return res.status(200).json({message: `contact ${req.params.id} deleted!`});
+    return res.status(204);
   } else {
-    return res.status(404).json({error: "Document not found."});
+    return res.status(404).json({error: result.error ||"Document not found."});
   }
 }
